Standardise to one animal management
