+++ src/europython_discord/bot.py
@@ -13,16 +13,10 @@ import discord
 from discord.ext import commands
 from pydantic import BaseModel
 
-from europython_discord.cat.cog import CatCog
-from europython_discord.cat.config import CatConfig
+from europython_discord.animals.cog import AnimalsCog
+from europython_discord.animals.config import AnimalsConfig
 from europython_discord.cogs.guild_statistics import GuildStatisticsCog, GuildStatisticsConfig
 from europython_discord.cogs.ping import PingCog
-from europython_discord.dog.cog import DogCog
-from europython_discord.dog.config import DogConfig
-from europython_discord.duck.cog import DuckCog
-from europython_discord.duck.config import DuckConfig
-from europython_discord.fox.cog import FoxCog
-from europython_discord.fox.config import FoxConfig
 from europython_discord.program_notifications.cog import ProgramNotificationsCog
 from europython_discord.program_notifications.config import ProgramNotificationsConfig
 from europython_discord.registration.cog import RegistrationCog
@@ -40,10 +34,7 @@ class Config(BaseModel):
     registration: RegistrationConfig
     program_notifications: ProgramNotificationsConfig
     guild_statistics: GuildStatisticsConfig
-    dog: DogConfig
-    cat: CatConfig
-    duck: DuckConfig
-    fox: FoxConfig
+    animals: AnimalsConfig
 
 
 async def run_bot(config: Config, auth_token: str) -> None:
@@ -56,10 +47,7 @@ async def run_bot(config: Config, auth_token: str) -> None:
 
     async with commands.Bot(intents=intents, command_prefix="$") as bot:
         await bot.add_cog(PingCog(bot))
-        await bot.add_cog(DogCog(bot, config.dog))
-        await bot.add_cog(CatCog(bot, config.cat))
-        await bot.add_cog(DuckCog(bot, config.duck))
-        await bot.add_cog(FoxCog(bot, config.fox))
+        await bot.add_cog(AnimalsCog(bot, config.animals))
         await bot.add_cog(RegistrationCog(bot, config.registration))
         await bot.add_cog(ProgramNotificationsCog(bot, config.program_notifications))
         await bot.add_cog(GuildStatisticsCog(bot, config.guild_statistics))
