--- src/europython_discord/cat/__init__.py
@@ -1,1 +1,0 @@
-# Cat module
