--- src/europython_discord/duck/__init__.py
@@ -1,1 +1,0 @@
-# Duck module
