--- src/europython_discord/fox/__init__.py
@@ -1,1 +1,0 @@
-# Fox module
