+++ test-config.toml
@@ -65,9 +65,11 @@ fast_mode = true
 [guild_statistics]
 required_role = "Organizers"
 
-[dog]
+[animals]
 channel_name = "animal-appreciation"
 cooldown_seconds = 10
+
+[animals.dog]
 error_messages = [
     "The dogs are on strike today! Try again later. 🐾🪧",
     "A wild error appeared! The dog got away... 🐕💨",
@@ -75,9 +77,7 @@ error_messages = [
     "404: Dog not found. Have you checked under the couch? 🛋️",
 ]
 
-[cat]
-channel_name = "animal-appreciation"
-cooldown_seconds = 10
+[animals.cat]
 error_messages = [
     "The cats are on strike today! Try again later. 🐾🪧",
     "A wild error appeared! The cat got away... 😿💨",
@@ -85,9 +85,7 @@ error_messages = [
     "404: Cat not found. Have you checked inside a cardboard box? 📦",
 ]
 
-[duck]
-channel_name = "animal-appreciation"
-cooldown_seconds = 10
+[animals.duck]
 error_messages = [
     "The ducks are on strike today! Try again later. 🐾🪧",
     "A wild error appeared! The duck got away... 🦆💨",
@@ -95,9 +93,7 @@ error_messages = [
     "Quack! The duck API seems to be having an off day.",
 ]
 
-[fox]
-channel_name = "animal-appreciation"
-cooldown_seconds = 10
+[animals.fox]
 error_messages = [
     "The foxes are on strike today! Try again later. 🐾🪧",
     "A wild error appeared! The fox got away... 🦊💨",
